Refactor label scaling and update Vico version
This commit introduces several changes to how label scaling is calculated and updates the Vico library version.

**Key changes:**

*   **Vico Version Update:**
    *   The Vico library version has been updated to `2.3.4`.
*   **Label Scaling Calculation:**
    *   The `visibleLabelsCount` property in `CartesianChart` and its composable counterpart now accepts a `Double` instead of an `Int`. This allows for more precise control over the number of visible labels, especially when dealing with fractional counts or dynamic scaling.
    *   The formula for calculating the label spacing scale factor in `CartesianChart.calculateLabelSpacingScaleFactor` has been adjusted. It now uses the full `layerDimensions.unscalablePadding` instead of a fraction of it, leading to more accurate spacing calculations.
*   **Padding in `LineCartesianLayer`:**
    *   The calculation for `unscalableStartPadding` and `unscalableEndPadding` in `LineCartesianLayer.HorizontalDimensions` no longer includes `maxPointSize.half`. This simplifies the padding logic and relies solely on the defined `layerPadding`.
*   **Range Coercion in `CartesianMeasuringContext`:**
    *   The `getFullXRange` and `getVisibleXRange` functions in `CartesianMeasuringContext` now coerce the calculated start and end values to be within the `ranges.minX` and `ranges.maxX` (or `fullRange.start` and `fullRange.endInclusive` respectively). This prevents the calculated ranges from extending beyond the actual data boundaries.
*   **Sample App Updates:**
    *   In `SaveableStateDemo.kt`:
        *   `Scroll.Absolute.xStable` has been changed to `Scroll.Absolute.x` for the initial scroll position.
        *   The `maxX` value for the chart data range has been increased from `80.9` to `85.9`.
        *   `visibleLabelsCount` for the top axis is now set as a `Double` (6.0).
    *   In `DynamicYRangeExample.kt`:
        *   `visibleLabelsCount` is now set as a `Double` (5.0).

+++ vico/compose/src/main/java/com/patrykandpatrick/vico/compose/cartesian/CartesianChart.kt
@@ -57,7 +57,7 @@ public fun rememberCartesianChart(
   fadingEdges: FadingEdges? = null,
   decorations: List<Decoration> = emptyList(),
   persistentMarkers: (CartesianChart.PersistentMarkerScope.(ExtraStore) -> Unit)? = null,
-  visibleLabelsCount: Int = 0,
+  visibleLabelsCount: Double = 0.0,
   getXStep: ((CartesianChartModel) -> Double) = { it.getXDeltaGcd() },
   onChartClick: ((List<Double>, Double?) -> Unit)? = null,
 ): CartesianChart {
